requirements.txt
binance-connector==3.12.0
python-dateutil>=2.8.2
python-dotenv>=1.0
requests>=2.32
pandas>=2.2
numpy>=2.0
pydantic>=1.10
prometheus-client>=0.17
optuna>=3